Conflicts:
	ConsomiTounsi/src/main/java/tn/esprit/spring/Controller/Forum/RepeatPaginator.java
	ConsomiTounsi/src/main/java/tn/esprit/spring/Repository/Forum/SujetRepository.java
	ConsomiTounsi/src/main/webapp/Template/template.xhtml
	ConsomiTounsi/src/main/webapp/WEB-INF/classes/META-INF/MANIFEST.MF
	ConsomiTounsi/src/main/webapp/WEB-INF/classes/META-INF/maven/tn.esprit.spring/ConsomiTounsi/pom.properties
	ConsomiTounsi/src/main/webapp/fourm/single-blog.xhtml

+++ ConsomiTounsi/src/main/java/tn/esprit/spring/Service/Panier/ICommande.java
@@ -20,8 +20,8 @@ public interface ICommande {
 	
 	public List< Commande> Commandeparcode( long code);
 	
-	public void PayerEnLigne(int idCommande,int iduser,String code);
+	public void PayerEnLigne(long idCommande,int iduser,String code);
 	
-	public void PayerPorteaPorte(int idCommande,int iduser,String code);
+	public void PayerPorteaPorte(long idCommande,int iduser,String code);
 
 }
